+++ Sources/NablaMessagingUI/Scenes/Inbox/Presenter/InboxPresenter.swift
@@ -1,0 +1,8 @@
+import Foundation
+import NablaMessagingCore
+
+// sourcery: AutoMockable
+protocol InboxPresenter: Presenter {
+    func userDidTapCreateConversation()
+    func userDidSelectConversation(_ conversation: Conversation)
+}