+++ Sources/NablaMessagingUI/Scenes/Inbox/UI/InboxViewContract.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol InboxViewContract: AnyObject {
+    func set(loading: Bool)
+    func display(error: String)
+}